Refactor MCP server implementation and update tooling.
Centralize server logic in `create_server.py` and refactor imports for better modularity. Update `Makefile` and `pyproject.toml` to reflect new script and command changes. Bump `ruff-pre-commit` version to v0.11.11 in `.pre-commit-config.yaml`.
